+++ quiz/view/index.php
@@ -15,7 +15,7 @@ include_once __DIR__ . "/../provider/provider.php";
         <a href="classement.php">lien vers le classement</a>
     </div>
     <div>
-        <a href="">lien vers la création d'un quiz</a>
+        <a href="creerQuiz.php">lien vers la création d'un quiz</a>
     </div>
     <p>Voici la liste des quiz disponibles:</p>
     <ul>
